webhook/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
httpx==0.26.0
pydantic==2.11.4
python-jose==3.3.0 
confluent-kafka==2.10.0